feat(add video & fix structer code): add video and model in create tour step2
