+++ 3d-virtual-nlu/src/components/admin/taskCreateTourList/Task3AddHotspot.tsx
@@ -1,32 +1,30 @@
-import * as THREE from "three";
-import React, { useState, useEffect, useRef } from "react";
+import { useState } from "react";
 import styles from "../../../styles/tasklistCT/task3.module.css";
-import { FaClock } from "react-icons/fa6";
-import { useDispatch, useSelector } from "react-redux";
+import { useSelector } from "react-redux";
 import { RootState } from "../../../redux/Store";
-import { FaHome } from "react-icons/fa";
-import axios from "axios";
-import UploadFile from "../UploadFile";
-import { useRaycaster } from "../../../hooks/useRaycaster";
 import TypeNavigation from "./HotspotNavigation";
 import { HotspotType } from "../../../redux/slices/HotspotSlice";
 import TypeInfomation from "./HotspotInformation";
 import TypeModel, { HotspotModelCreateRequest } from "./HotspotModel";
-import TypeMedia, { CornerPoint, HotspotMediaCreateRequest } from "./HotspotMedia";
+import TypeMedia, {
+  CornerPoint,
+  HotspotMediaCreateRequest,
+} from "./HotspotMedia";
+import ConfigIcon from "../ConfigIcon";
 
 interface Task3Props {
   assignable: boolean;
   setAssignable: (value: boolean) => void;
   hotspotModels: HotspotModelCreateRequest[];
-  setHotspotModels: (value: HotspotModelCreateRequest[]) => void;
+  // setHotspotModels: (value: HotspotModelCreateRequest[]) => void;
   chooseCornerMediaPoint: boolean;
   setChooseCornerMediaPoint: (value: boolean) => void;
   currentPoints: [number, number, number][]; // mesh đang chọn
   setCurrentPoints: (val: any) => void;
-  videoMeshes: HotspotMediaCreateRequest[]; // danh sách mesh đã xong
-  setVideoMeshes: (val: any) => void;
-  cornerPointes: CornerPoint[]; // danh sách mesh đã xong
-  setCornerPointes: (val: any) => void;
+  // videoMeshes: HotspotMediaCreateRequest[]; // danh sách mesh đã xong
+  // setVideoMeshes: (val: any) => void;
+  // cornerPointes: CornerPoint[]; // danh sách mesh đã xong
+  // setCornerPointes: (val: any) => void;
   currentHotspotType: HotspotType | null;
   setCurrentHotspotType: (value: HotspotType) => void;
 }
@@ -36,17 +34,13 @@ const Task3 = ({
   assignable,
   setAssignable,
   hotspotModels,
-  setHotspotModels,
+  // setHotspotModels,
   chooseCornerMediaPoint,
   setChooseCornerMediaPoint,
-  videoMeshes,
   currentPoints,
   setCurrentPoints,
-  setVideoMeshes,
-  currentHotspotType,
+  // setVideoMeshes,
   setCurrentHotspotType,
-  cornerPointes,
-  setCornerPointes,
 }: Task3Props) => {
   const [openTypeIndex, setOpenTypeIndex] = useState<number | null>(1); // State để lưu index của type đang mở
   const hotspotType = useSelector(
@@ -71,12 +65,13 @@ const Task3 = ({
           ))}
         </select>
       </div>
+      {/* // setup icon */}
+      <ConfigIcon />
       <TypeNavigation
         isOpenTypeNavigation={openTypeIndex == 1}
         setAssignable={setAssignable}
         setCurrentHotspotType={setCurrentHotspotType}
       />
-
       <TypeInfomation
         isOpenTypeInfomation={openTypeIndex == 2}
         setAssignable={setAssignable}
@@ -85,20 +80,16 @@ const Task3 = ({
       <TypeMedia
         isOpenTypeMedia={openTypeIndex == 3}
         currentPoints={currentPoints}
-        videoMeshes={videoMeshes}
         setCurrentPoints={setCurrentPoints}
-        setVideoMeshes={setVideoMeshes}
         chooseCornerMediaPoint={chooseCornerMediaPoint}
         setChooseCornerMediaPoint={setChooseCornerMediaPoint}
-        // cornerPointes={cornerPointes}
-        // setCornerPointes={setCornerPointes}
         setAssignable={setAssignable}
         setCurrentHotspotType={setCurrentHotspotType}
       />
       <TypeModel
         isOpenTypeModel={openTypeIndex == 4}
         hotspotModels={hotspotModels}
-        setHotspotModels={setHotspotModels}
+        // setHotspotModels={setHotspotModels}
         assignable={assignable}
         setAssignable={setAssignable}
         setCurrentHotspotType={setCurrentHotspotType}
